backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
qdrant-client>=1.7.1
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
python-multipart>=0.0.6
tqdm>=4.66.1
sentence-transformers